requirements.txt
Django==1.6.8
Markdown==2.5.1
South==1.0.1
djangorestframework==2.4.4
gnureadline==6.3.3
ipython==2.3.0
requests==2.4.3
wsgiref==0.1.2
